--- web/models.go
@@ -1,70 +1,0 @@
-// Copyright 2020 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package web
-
-import (
-	"time"
-
-	"github.com/cloudspannerecosystem/harbourbridge/internal"
-)
-
-// Driver config is used for direct database connection
-type DriverConfig struct {
-	Driver   string `json:"Driver"`
-	Host     string `json:"Host"`
-	Port     string `json:"Port"`
-	Database string `json:"Database"`
-	User     string `json:"User"`
-	Password string `json:"Password"`
-}
-
-// Dump config is used for dump file approach
-type DumpConfig struct {
-	Driver   string `json:"Driver"`
-	FilePath string `json:"Path"`
-}
-
-// Metadata of session file, which contains app.conv in
-// JSON format
-type Session struct {
-	Driver    string    `json:"driver"`
-	FilePath  string    `json:"path"`
-	FileName  string    `json:"fileName"`
-	CreatedAt time.Time `json:"createdAt"`
-}
-
-// Type and issue
-type typeIssue struct {
-	T     string
-	Issue internal.SchemaIssue
-	Brief string
-}
-
-// Actions to be performed on a column.
-// (1) Removed: true/false
-// (2) Rename: New name or empty string
-// (3) PK: "ADDED", "REMOVED" or ""
-// (4) NotNull: "ADDED", "REMOVED" or ""
-// (5) ToType: New type or empty string
-type updateCol struct {
-	Removed bool   `json:"Removed"`
-	Rename  string `json:"Rename"`
-	PK      string `json:"PK"`
-	NotNull string `json:"NotNull"`
-	ToType  string `json:"ToType"`
-}
-type updateTable struct {
-	UpdateCols map[string]updateCol `json:"UpdateCols"`
-}
