+++ web/web_test.go
@@ -66,7 +66,7 @@ func TestGetTypeMapPostgres(t *testing.T) {
 			typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief}},
 		"bigserial": []typeIssue{
 			typeIssue{T: ddl.Int64, Brief: internal.IssueDB[internal.Serial].Brief},
-			typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief}},
+			typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief + ", " + internal.IssueDB[internal.Serial].Brief}},
 		"bpchar": []typeIssue{
 			typeIssue{T: ddl.Bytes},
 			typeIssue{T: ddl.String}},
@@ -93,7 +93,7 @@ func TestGetTypeMapPostgres(t *testing.T) {
 			typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief}},
 		"serial": []typeIssue{
 			typeIssue{T: ddl.Int64, Brief: internal.IssueDB[internal.Serial].Brief},
-			typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief}},
+			typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief + ", " + internal.IssueDB[internal.Serial].Brief}},
 		"text": []typeIssue{
 			typeIssue{T: ddl.Bytes},
 			typeIssue{T: ddl.String}},
@@ -176,11 +176,11 @@ func TestSetTypeMapTableLevelPostgres(t *testing.T) {
 				"c": []internal.SchemaIssue{internal.Widened},
 				"e": []internal.SchemaIssue{internal.Widened},
 				"f": []internal.SchemaIssue{internal.Widened},
-				"g": []internal.SchemaIssue{internal.Widened},
+				"g": []internal.SchemaIssue{internal.Widened, internal.Serial},
 				"j": []internal.SchemaIssue{internal.Widened},
 				"k": []internal.SchemaIssue{internal.Widened},
 				"l": []internal.SchemaIssue{internal.Widened},
-				"m": []internal.SchemaIssue{internal.Widened},
+				"m": []internal.SchemaIssue{internal.Widened, internal.Serial},
 				"o": []internal.SchemaIssue{internal.Widened},
 				"p": []internal.SchemaIssue{internal.Widened},
 			},
@@ -538,11 +538,11 @@ func TestSetTypeMapGlobalLevelPostgres(t *testing.T) {
 				"c": []internal.SchemaIssue{internal.Widened},
 				"e": []internal.SchemaIssue{internal.Widened},
 				"f": []internal.SchemaIssue{internal.Widened},
-				"g": []internal.SchemaIssue{internal.Widened},
+				"g": []internal.SchemaIssue{internal.Widened, internal.Serial},
 				"j": []internal.SchemaIssue{internal.Widened},
 				"k": []internal.SchemaIssue{internal.Widened},
 				"l": []internal.SchemaIssue{internal.Widened},
-				"m": []internal.SchemaIssue{internal.Widened},
+				"m": []internal.SchemaIssue{internal.Widened, internal.Serial},
 				"o": []internal.SchemaIssue{internal.Widened},
 				"p": []internal.SchemaIssue{internal.Widened},
 			},
